src/main/java/com/example/recetas_back/util/JWTUtil.java
package com.example.recetas_back.util;

import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;
import io.jsonwebtoken.security.Keys;

import org.springframework.stereotype.Component;

import java.security.Key;
import java.util.Date;

import javax.crypto.SecretKey;

@Component
public class JWTUtil {

    private final SecretKey key = Keys.hmacShaKeyFor("myFixedSecretKeyWhichIsAtLeast32Characters".getBytes());

    public String generateToken(String username, String role) {
        return Jwts.builder()
                .setSubject(username)
                .claim("role", role) // Incluye el rol como un claim
                .setIssuedAt(new Date())
                .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 * 60 * 10)) // 10 horas
                .signWith(key) // Usa tu clave secreta
                .compact();
    }

    public String extractRole(String token) {
        Claims claims = Jwts.parser()
                .setSigningKey(key) // Usa tu clave secreta
                .parseClaimsJws(token)
                .getBody();
        return claims.get("role", String.class); // Extrae el rol
    }

    public boolean validateToken(String token, String username) {
        return username.equals(extractUsername(token)) && !isTokenExpired(token);
    }

    public String extractUsername(String token) {
        return getClaims(token).getSubject();
    }

    private Claims getClaims(String token) {
        return Jwts.parserBuilder()
                .setSigningKey(key)
                .build()
                .parseClaimsJws(token)
                .getBody();
    }

    private boolean isTokenExpired(String token) {
        return getClaims(token).getExpiration().before(new Date());
    }
}
